config/requirement.txt
python-jenkins==1.8.1
PyYAML==6.0.1